backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
PyJWT==2.8.1
bcrypt==4.1.1
gunicorn==21.2.0
requests==2.31.0
Werkzeug==3.0.1
click==8.1.7
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
